Support `asakusafw.sdk.*` convention on Gradle.
The following convention property is also now available:
* `asakusafw.m3bp.version`
  * Asakusa on M3BP libraries version.
