+++ gradle/src/main/groovy/com/asakusafw/m3bp/gradle/plugins/internal/AsakusaM3bpSdkPlugin.groovy
@@ -28,16 +28,16 @@ import com.asakusafw.gradle.tasks.internal.ResolutionUtils
 
 /**
  * A Gradle sub plug-in for Asakusa on M3BP SDK.
+ * @since 0.1.0
+ * @version 0.3.0
+ * @see AsakusaM3bpSdkBasePlugin
  */
 class AsakusaM3bpSdkPlugin implements Plugin<Project> {
 
+    /**
+     * The compile batch application task name.
+     */
     public static final String TASK_COMPILE = 'm3bpCompileBatchapps'
-
-    private static final Map<String, String> REDIRECT = [
-            'com.asakusafw.runtime.core.BatchContext' : 'com.asakusafw.m3bp.custom.M3bpBatchContext',
-            'com.asakusafw.runtime.core.Report' : 'com.asakusafw.bridge.api.Report',
-            'com.asakusafw.runtime.directio.api.DirectIo' : 'com.asakusafw.bridge.directio.api.DirectIo',
-    ]
 
     private Project project
 
@@ -47,71 +47,10 @@ class AsakusaM3bpSdkPlugin implements Plugin<Project> {
     void apply(Project project) {
         this.project = project
 
-        project.apply plugin: 'asakusafw-sdk'
-        project.apply plugin: AsakusaM3bpBasePlugin
-        extension = AsakusaSdkPlugin.get(project).extensions.create('m3bp', AsakusafwCompilerExtension)
+        project.apply plugin: AsakusaM3bpSdkBasePlugin
+        this.extension = AsakusaM3bpSdkBasePlugin.get(project)
 
-        configureExtension()
-        configureConfigurations()
         defineTasks()
-    }
-
-    private void configureExtension() {
-        AsakusafwPluginConvention sdk = AsakusaSdkPlugin.get(project)
-        extension.conventionMapping.with {
-            outputDirectory = { project.relativePath(new File(project.buildDir, 'm3bp-batchapps')) }
-            batchIdPrefix = { (String) 'm3bp.' }
-            failOnError = { true }
-        }
-        REDIRECT.each { k, v ->
-            extension.compilerProperties.put((String) "redirector.rule.${k}", v)
-        }
-        extension.compilerProperties.put('javac.version', { sdk.javac.sourceCompatibility.toString() })
-    }
-
-    private void configureConfigurations() {
-        project.configurations {
-            asakusaM3bpCommon {
-                description 'Common libraries of Asakusa DSL Compiler for M3BP'
-                exclude group: 'asm', module: 'asm'
-            }
-            asakusaM3bpCompiler {
-                description 'Full classpath of Asakusa DSL Compiler for M3BP'
-                extendsFrom project.configurations.compile
-                extendsFrom project.configurations.asakusaM3bpCommon
-            }
-            asakusaM3bpTestkit {
-                description 'Asakusa DSL testkit classpath for M3BP'
-                extendsFrom project.configurations.asakusaM3bpCommon
-                exclude group: 'com.asakusafw', module: 'asakusa-test-mapreduce'
-            }
-        }
-        PluginUtils.afterEvaluate(project) {
-            AsakusaM3bpBaseExtension base = AsakusaM3bpBasePlugin.get(project)
-            AsakusafwPluginConvention sdk = AsakusaSdkPlugin.get(project)
-            project.dependencies {
-                asakusaM3bpCommon "com.asakusafw.m3bp.compiler:asakusa-m3bp-compiler-core:${base.featureVersion}"
-                asakusaM3bpCommon "com.asakusafw.dag.compiler:asakusa-dag-compiler-extension-directio:${base.langVersion}"
-                asakusaM3bpCommon "com.asakusafw.dag.compiler:asakusa-dag-compiler-extension-windgate:${base.langVersion}"
-                asakusaM3bpCommon "com.asakusafw.lang.compiler:asakusa-compiler-cli:${base.langVersion}"
-                asakusaM3bpCommon "com.asakusafw.lang.compiler:asakusa-compiler-extension-redirector:${base.langVersion}"
-                asakusaM3bpCommon "com.asakusafw.lang.compiler:asakusa-compiler-extension-yaess:${base.langVersion}"
-                asakusaM3bpCommon "com.asakusafw.lang.compiler:asakusa-compiler-extension-directio:${base.langVersion}"
-                asakusaM3bpCommon "com.asakusafw.lang.compiler:asakusa-compiler-extension-hive:${base.langVersion}"
-                asakusaM3bpCommon "com.asakusafw.lang.compiler:asakusa-compiler-extension-windgate:${base.langVersion}"
-                asakusaM3bpCommon "com.asakusafw:simple-graph:${sdk.asakusafwVersion}"
-                asakusaM3bpCommon "com.asakusafw:java-dom:${sdk.asakusafwVersion}"
-
-                asakusaM3bpCompiler "com.asakusafw:asakusa-dsl-vocabulary:${sdk.asakusafwVersion}"
-                asakusaM3bpCompiler "com.asakusafw:asakusa-runtime:${sdk.asakusafwVersion}"
-                asakusaM3bpCompiler "com.asakusafw:asakusa-yaess-core:${sdk.asakusafwVersion}"
-                asakusaM3bpCompiler "com.asakusafw:asakusa-directio-vocabulary:${sdk.asakusafwVersion}"
-                asakusaM3bpCompiler "com.asakusafw:asakusa-windgate-vocabulary:${sdk.asakusafwVersion}"
-
-                asakusaM3bpTestkit "com.asakusafw.m3bp.compiler:asakusa-m3bp-compiler-test-adapter:${base.featureVersion}"
-                asakusaM3bpTestkit "com.asakusafw.m3bp.bridge:asakusa-m3bp-assembly:${base.featureVersion}"
-            }
-        }
     }
 
     private void defineTasks() {
